Separate out ocean sea depth
