+++ src/unity/Assets/Crest/Scripts/LodData/LodDataAnimatedWaves.cs
@@ -13,19 +13,12 @@ namespace Crest
     ///  * The textures from this LodData are passed to the ocean material when the surface is drawn
     ///  * Foam and Dynamic Waves LodDatas add their results into this LodData. The dynamic waves piggy back off the combine
     ///    pass. Both piggy back off the assignment to the ocean material.
-    ///  * This LodData currently also takes care of rendering the depth of the ocean sea floor. This could/should be moved out.
+    ///  * The LodDataSeaFloorDepth sits on this same GameObject and borrows the camera. This could be a model for the other sim types..
     /// </summary>
     public class LodDataAnimatedWaves : LodData
     {
         // debug use
         public static bool _shapeCombinePass = true;
-
-        Material _matOceanDepth;
-        RenderTexture _rtOceanDepth;
-        CommandBuffer _bufOceanDepth = null;
-        bool _oceanDepthRenderersDirty = true;
-        /// <summary>Called when one or more objects that will render into depth are created, so that all objects are registered.</summary>
-        public void OnOceanDepthRenderersChanged() { _oceanDepthRenderersDirty = true; }
 
         public override SimSettingsBase CreateDefaultSettings() { return null; }
         public override void UseSettings(SimSettingsBase settings) {}
@@ -46,9 +39,6 @@ namespace Crest
 
         void Start()
         {
-            Cam.depthTextureMode = DepthTextureMode.None;
-
-            _matOceanDepth = new Material(Shader.Find("Ocean/Ocean Depth"));
             _combineMaterial = new Material(Shader.Find("Ocean/Shape/Combine"));
 
             // create shader param IDs for each LOD once on start to avoid creating garbage each frame.
@@ -69,12 +59,6 @@ namespace Crest
             if (LodIndex == 0)
             {
                 UpdateCmdBufShapeCombine();
-            }
-
-            if (_oceanDepthRenderersDirty)
-            {
-                UpdateCmdBufOceanFloorDepth();
-                _oceanDepthRenderersDirty = false;
             }
         }
 
@@ -115,65 +99,6 @@ namespace Crest
             }
         }
 
-        // The command buffer populates the LODs with ocean depth data. It submits any objects with a RenderOceanDepth component attached.
-        // It's stateless - the textures don't have to be managed across frames/scale changes
-        void UpdateCmdBufOceanFloorDepth()
-        {
-            var objs = FindObjectsOfType<RenderOceanDepth>();
-
-            // if there is nothing in the scene tagged up for depth rendering then there is no depth rendering required
-            if (objs.Length < 1)
-            {
-                if (_bufOceanDepth != null)
-                {
-                    _bufOceanDepth.Clear();
-                }
-
-                return;
-            }
-
-            if (!_rtOceanDepth)
-            {
-                _rtOceanDepth = new RenderTexture(Cam.targetTexture.width, Cam.targetTexture.height, 0);
-                _rtOceanDepth.name = gameObject.name + "_oceanDepth";
-                _rtOceanDepth.format = RenderTextureFormat.RHalf;
-                _rtOceanDepth.useMipMap = false;
-                _rtOceanDepth.anisoLevel = 0;
-            }
-
-            if (_bufOceanDepth == null)
-            {
-                _bufOceanDepth = new CommandBuffer();
-                Cam.AddCommandBuffer(CameraEvent.BeforeForwardOpaque, _bufOceanDepth);
-                _bufOceanDepth.name = "Ocean Depth";
-            }
-
-            _bufOceanDepth.Clear();
-
-            _bufOceanDepth.SetRenderTarget( _rtOceanDepth );
-            _bufOceanDepth.ClearRenderTarget( false, true, Color.red * 10000f );
-
-            foreach (var obj in objs)
-            {
-                if (!obj.enabled)
-                    continue;
-
-                var r = obj.GetComponent<Renderer>();
-                if (r == null)
-                {
-                    Debug.LogError("GameObject '" + obj.gameObject.name + "' must have a renderer component attached. Unity Terrain objects are not supported - these must be captured by an Ocean Depth Cache.", obj);
-                }
-                else if (obj.transform.parent != null && obj.transform.parent.GetComponent<OceanDepthCache>() != null)
-                {
-                    _bufOceanDepth.DrawRenderer(r, r.material);
-                }
-                else
-                {
-                    _bufOceanDepth.DrawRenderer(r, _matOceanDepth);
-                }
-            }
-        }
-
         /// <summary>
         /// Additively combine shape from biggest LODs to smallest LOD. Executed once per frame - attached to the LOD0 camera which
         /// is the last LOD camera to render.
@@ -210,12 +135,6 @@ namespace Crest
 
         void RemoveCommandBuffers()
         {
-            if( _bufOceanDepth != null )
-            {
-                Cam.RemoveCommandBuffer(CameraEvent.BeforeForwardOpaque, _bufOceanDepth);
-                _bufOceanDepth = null;
-            }
-
             if (_bufCombineShapes != null)
             {
                 Cam.RemoveCommandBuffer(CameraEvent.AfterEverything, _bufCombineShapes);
@@ -276,9 +195,9 @@ namespace Crest
                 properties.SetTexture(_paramsDisplacementsSampler[shapeSlot], Cam.targetTexture);
             }
 
-            if (_rtOceanDepth != null)
+            if (LDSD.RTOceanDepth != null)
             {
-                properties.SetTexture(_paramsOceanDepthSampler[shapeSlot], _rtOceanDepth);
+                properties.SetTexture(_paramsOceanDepthSampler[shapeSlot], LDSD.RTOceanDepth);
             }
 
             // need to blend out shape if this is the largest lod, and the ocean might get scaled down later (so the largest lod will disappear)
@@ -294,5 +213,8 @@ namespace Crest
         ReadbackDisplacementsForCollision _collReadback;
         public ReadbackDisplacementsForCollision CollReadback
         { get { return _collReadback ?? (_collReadback = GetComponent<ReadbackDisplacementsForCollision>()); } }
+
+        LodDataSeaFloorDepth _ldsd;
+        public LodDataSeaFloorDepth LDSD { get { return _ldsd ?? (_ldsd = GetComponent<LodDataSeaFloorDepth>()); } }
     }
 }
