+++ src/unity/Assets/Crest/Scripts/LodData/LodDataSeaFloorDepth.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8761dee849d4d14ea059a38942940ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
