--- src/unity/Assets/Crest/Shaders/Simulation/Resources/ShapeSimAddToDisps.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 2a0b8abf8f1036044abe2882672a745d
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
